Move activate-dropped-tab under preview dir.
